Add tool CLI subcommand.
