+++ quickwit/quickwit-cli/src/index.rs
@@ -17,14 +17,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use std::collections::{HashSet, VecDeque};
+use std::collections::VecDeque;
 use std::fmt::Display;
 use std::io::{stdout, Stdout, Write};
 use std::path::PathBuf;
 use std::str::FromStr;
-use std::sync::Arc;
 use std::time::{Duration, Instant};
-use std::{env, fmt, io};
+use std::{fmt, io};
 
 use anyhow::{bail, Context};
 use bytes::Bytes;
@@ -32,51 +31,38 @@ use clap::{arg, ArgMatches, Command};
 use colored::{ColoredString, Colorize};
 use humantime::format_duration;
 use itertools::Itertools;
-use quickwit_actors::{ActorExitStatus, ActorHandle, ObservationType, Universe};
-use quickwit_cluster::create_fake_cluster_for_cli;
+use quickwit_actors::{ActorHandle, ObservationType};
 use quickwit_common::uri::Uri;
-use quickwit_common::{GREEN_COLOR, RED_COLOR};
-use quickwit_config::service::QuickwitService;
-use quickwit_config::{
-    ConfigFormat, IndexConfig, IndexerConfig, SourceConfig, SourceParams, TransformConfig,
-    VecSourceParams, CLI_INGEST_SOURCE_ID,
-};
-use quickwit_core::{clear_cache_directory, IndexService};
-use quickwit_indexing::actors::{IndexingService, MergePipeline, MergePipelineId};
-use quickwit_indexing::models::{
-    DetachIndexingPipeline, DetachMergePipeline, IndexingStatistics, SpawnPipeline,
-};
+use quickwit_common::GREEN_COLOR;
+use quickwit_config::{ConfigFormat, IndexConfig};
+use quickwit_indexing::models::IndexingStatistics;
 use quickwit_indexing::IndexingPipeline;
-use quickwit_metastore::{quickwit_metastore_uri_resolver, IndexMetadata, Split};
+use quickwit_metastore::{IndexMetadata, Split};
 use quickwit_proto::SortOrder;
 use quickwit_rest_client::models::IngestSource;
 use quickwit_rest_client::rest_client::{QuickwitClient, Transport};
 use quickwit_search::SearchResponseRest;
 use quickwit_serve::{ListSplitsQueryParams, SearchRequestQueryString, SortByField};
-use quickwit_storage::{load_file, quickwit_storage_uri_resolver};
+use quickwit_storage::load_file;
 use quickwit_telemetry::payload::TelemetryEvent;
 use reqwest::Url;
 use tabled::object::{Columns, Segment};
 use tabled::{Alignment, Concat, Format, Modify, Panel, Rotate, Style, Table, Tabled};
 use thousands::Separable;
-use tracing::{debug, info, Level};
+use tracing::{debug, Level};
 
 use crate::stats::{mean, percentile, std_deviation};
-use crate::{
-    cluster_endpoint_arg, config_cli_arg, load_quickwit_config, make_table,
-    parse_duration_with_unit, prompt_confirmation, run_index_checklist, start_actor_runtimes,
-    THROUGHPUT_WINDOW_SIZE,
-};
+use crate::{cluster_endpoint_arg, make_table, prompt_confirmation, THROUGHPUT_WINDOW_SIZE};
 
 pub fn build_index_command<'a>() -> Command<'a> {
     Command::new("index")
         .about("Manages indexes: creates, deletes, ingests, searches, describes...")
+        .arg(cluster_endpoint_arg())
         .subcommand(
             Command::new("create")
                 .display_order(1)
                 .about("Creates an index from an index config file.")
                 .args(&[
-                    cluster_endpoint_arg(),
                     arg!(--"index-config" <INDEX_CONFIG> "Location of the index config file."),
                     arg!(--overwrite "Overwrites pre-existing index. This will delete all existing data stored at `index-uri` before creating a new index.")
                         .required(false),
@@ -89,7 +75,6 @@ pub fn build_index_command<'a>() -> Command<'a> {
                 .about("Clears an index: deletes all splits and resets checkpoint.")
                 .long_about("Deletes all its splits and resets its checkpoint. This operation is destructive and cannot be undone, proceed with caution.")
                 .args(&[
-                    cluster_endpoint_arg(),
                     arg!(--index <INDEX> "Index ID")
                         .display_order(1),
                 ])
@@ -101,7 +86,6 @@ pub fn build_index_command<'a>() -> Command<'a> {
                 .about("Deletes an index.")
                 .long_about("Deletes an index. This operation is destructive and cannot be undone, proceed with caution.")
                 .args(&[
-                    cluster_endpoint_arg(),
                     arg!(--index <INDEX> "ID of the target index")
                         .display_order(1),
                     arg!(--"dry-run" "Executes the command in dry run mode and only displays the list of splits candidates for deletion.")
@@ -114,7 +98,6 @@ pub fn build_index_command<'a>() -> Command<'a> {
                 .about("Displays descriptive statistics of an index.")
                 .long_about("Displays descriptive statistics of an index. Displayed statistics are: number of published splits, number of documents, splits min/max timestamps, size of splits.")
                 .args(&[
-                    cluster_endpoint_arg(),
                     arg!(--index <INDEX> "ID of the target index")
                         .display_order(1),
                 ])
@@ -132,7 +115,6 @@ pub fn build_index_command<'a>() -> Command<'a> {
                 .about("Ingest NDJSON documents with the ingest API.")
                 .long_about("Reads NDJSON documents from a file or streamed from stdin and sends them into ingest API.")
                 .args(&[
-                    cluster_endpoint_arg(),
                     arg!(--index <INDEX> "ID of the target index")
                         .display_order(1),
                     arg!(--"input-path" <INPUT_PATH> "Location of the input file.")
@@ -144,7 +126,6 @@ pub fn build_index_command<'a>() -> Command<'a> {
                 .display_order(7)
                 .about("Searches an index.")
                 .args(&[
-                    cluster_endpoint_arg(),
                     arg!(--index <INDEX> "ID of the target index")
                         .display_order(1),
                     arg!(--query <QUERY> "Query expressed in natural query language ((barack AND obama) OR \"president of united states\"). Learn more on https://quickwit.io/docs/reference/search-language."),
@@ -168,51 +149,6 @@ pub fn build_index_command<'a>() -> Command<'a> {
                         .required(false),
                     arg!(--"sort-by-score" "Setting this flag calculates and sorts documents by their BM25 score.")
                         .required(false),
-                ])
-            )
-        .subcommand(
-            Command::new("local-ingest")
-                .display_order(10)
-                .about("Indexes NDJSON documents locally.")
-                .long_about("Local ingest starts a quickwit isntance locally and directly indexes NDJSON documents from a file or from stdin.")
-                .args(&[
-                    config_cli_arg(),
-                    arg!(--index <INDEX> "ID of the target index")
-                        .display_order(1),
-                    arg!(--"input-path" <INPUT_PATH> "Location of the input file.")
-                        .required(false),
-                    arg!(--overwrite "Overwrites pre-existing index.")
-                        .required(false),
-                    arg!(--"transform-script" <SCRIPT> "VRL program to transform docs before ingesting.")
-                        .required(false),
-                    arg!(--"keep-cache" "Does not clear local cache directory upon completion.")
-                        .required(false),
-                ])
-            )
-        .subcommand(
-            Command::new("gc")
-                .display_order(10)
-                .about("Garbage collects stale staged splits and splits marked for deletion.")
-                .args(&[
-                    config_cli_arg(),
-                    arg!(--index <INDEX> "ID of the target index")
-                        .display_order(1),
-                    arg!(--"grace-period" <GRACE_PERIOD> "Threshold period after which stale staged splits are garbage collected.")
-                        .default_value("1h")
-                        .required(false),
-                    arg!(--"dry-run" "Executes the command in dry run mode and only displays the list of splits candidates for garbage collection.")
-                        .required(false),
-                ])
-            )
-        .subcommand(
-            Command::new("merge")
-                .display_order(10)
-                .about("Merges all the splits for a given Node ID, index ID, source ID.")
-                .args(&[
-                    config_cli_arg(),
-                    arg!(--index <INDEX> "ID of the target index.")
-                        .display_order(1),
-                    arg!(--source <SOURCE_ID> "ID of the target source."),
                 ])
             )
         .arg_required_else_help(true)
@@ -247,16 +183,6 @@ pub struct IngestDocsArgs {
 }
 
 #[derive(Debug, Eq, PartialEq)]
-pub struct LocalIngestDocsArgs {
-    pub config_uri: Uri,
-    pub index_id: String,
-    pub input_path_opt: Option<PathBuf>,
-    pub overwrite: bool,
-    pub vrl_script: Option<String>,
-    pub clear_cache: bool,
-}
-
-#[derive(Debug, Eq, PartialEq)]
 pub struct SearchIndexArgs {
     pub cluster_endpoint: Url,
     pub index_id: String,
@@ -280,21 +206,6 @@ pub struct DeleteIndexArgs {
 }
 
 #[derive(Debug, Eq, PartialEq)]
-pub struct GarbageCollectIndexArgs {
-    pub config_uri: Uri,
-    pub index_id: String,
-    pub grace_period: Duration,
-    pub dry_run: bool,
-}
-
-#[derive(Debug, Eq, PartialEq)]
-pub struct MergeArgs {
-    pub config_uri: Uri,
-    pub index_id: String,
-    pub source_id: String,
-}
-
-#[derive(Debug, Eq, PartialEq)]
 pub struct ListIndexesArgs {
     pub cluster_endpoint: Url,
 }
@@ -305,11 +216,8 @@ pub enum IndexCliCommand {
     Create(CreateIndexArgs),
     Delete(DeleteIndexArgs),
     Describe(DescribeIndexArgs),
-    GarbageCollect(GarbageCollectIndexArgs),
     Ingest(IngestDocsArgs),
-    LocalIngest(LocalIngestDocsArgs),
     List(ListIndexesArgs),
-    Merge(MergeArgs),
     Search(SearchIndexArgs),
 }
 
@@ -330,11 +238,8 @@ impl IndexCliCommand {
             "create" => Self::parse_create_args(submatches),
             "delete" => Self::parse_delete_args(submatches),
             "describe" => Self::parse_describe_args(submatches),
-            "gc" => Self::parse_garbage_collect_args(submatches),
             "ingest" => Self::parse_ingest_args(submatches),
-            "local-ingest" => Self::parse_local_ingest_args(submatches),
             "list" => Self::parse_list_args(submatches),
-            "merge" => Self::parse_merge_args(submatches),
             "search" => Self::parse_search_args(submatches),
             _ => bail!("Index subcommand `{}` is not implemented.", subcommand),
         }
@@ -424,38 +329,6 @@ impl IndexCliCommand {
         }))
     }
 
-    fn parse_local_ingest_args(matches: &ArgMatches) -> anyhow::Result<Self> {
-        let config_uri = matches
-            .value_of("config")
-            .map(Uri::from_str)
-            .expect("`config` is a required arg.")?;
-        let index_id = matches
-            .value_of("index")
-            .expect("`index` is a required arg.")
-            .to_string();
-        let input_path_opt = if let Some(input_path) = matches.value_of("input-path") {
-            Uri::from_str(input_path)?
-                .filepath()
-                .map(|path| path.to_path_buf())
-        } else {
-            None
-        };
-        let overwrite = matches.is_present("overwrite");
-        let vrl_script = matches
-            .value_of("transform-script")
-            .map(|source| source.to_string());
-        let clear_cache = !matches.is_present("keep-cache");
-
-        Ok(Self::LocalIngest(LocalIngestDocsArgs {
-            config_uri,
-            index_id,
-            input_path_opt,
-            overwrite,
-            vrl_script,
-            clear_cache,
-        }))
-    }
-
     fn parse_search_args(matches: &ArgMatches) -> anyhow::Result<Self> {
         let index_id = matches
             .value_of("index")
@@ -505,48 +378,6 @@ impl IndexCliCommand {
         }))
     }
 
-    fn parse_merge_args(matches: &ArgMatches) -> anyhow::Result<Self> {
-        let config_uri = matches
-            .value_of("config")
-            .map(Uri::from_str)
-            .expect("`config` is a required arg.")?;
-        let index_id = matches
-            .value_of("index")
-            .context("'index-id' is a required arg.")?
-            .to_string();
-        let source_id = matches
-            .value_of("source")
-            .context("'source-id' is a required arg.")?
-            .to_string();
-        Ok(Self::Merge(MergeArgs {
-            index_id,
-            source_id,
-            config_uri,
-        }))
-    }
-
-    fn parse_garbage_collect_args(matches: &ArgMatches) -> anyhow::Result<Self> {
-        let config_uri = matches
-            .value_of("config")
-            .map(Uri::from_str)
-            .expect("`config` is a required arg.")?;
-        let index_id = matches
-            .value_of("index")
-            .expect("`index` is a required arg.")
-            .to_string();
-        let grace_period = matches
-            .value_of("grace-period")
-            .map(parse_duration_with_unit)
-            .expect("`grace-period` should have a default value.")?;
-        let dry_run = matches.is_present("dry-run");
-        Ok(Self::GarbageCollect(GarbageCollectIndexArgs {
-            index_id,
-            grace_period,
-            dry_run,
-            config_uri,
-        }))
-    }
-
     fn parse_delete_args(matches: &ArgMatches) -> anyhow::Result<Self> {
         let cluster_endpoint = matches
             .value_of("endpoint")
@@ -572,11 +403,8 @@ impl IndexCliCommand {
             Self::Create(args) => create_index_cli(args).await,
             Self::Delete(args) => delete_index_cli(args).await,
             Self::Describe(args) => describe_index_cli(args).await,
-            Self::GarbageCollect(args) => garbage_collect_index_cli(args).await,
             Self::Ingest(args) => ingest_docs_cli(args).await,
-            Self::LocalIngest(args) => local_ingest_docs_cli(args).await,
             Self::List(args) => list_index_cli(args).await,
-            Self::Merge(args) => merge_cli(args).await,
             Self::Search(args) => search_index_cli(args).await,
         }
     }
@@ -918,118 +746,6 @@ pub async fn ingest_docs_cli(args: IngestDocsArgs) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub async fn local_ingest_docs_cli(args: LocalIngestDocsArgs) -> anyhow::Result<()> {
-    debug!(args=?args, "local-ingest-docs");
-    println!("❯ Ingesting documents locally...");
-    quickwit_telemetry::send_telemetry_event(TelemetryEvent::Ingest).await;
-
-    let config = load_quickwit_config(&args.config_uri).await?;
-
-    let source_params = if let Some(filepath) = args.input_path_opt.as_ref() {
-        SourceParams::file(filepath)
-    } else {
-        SourceParams::stdin()
-    };
-    let transform_config = args
-        .vrl_script
-        .map(|vrl_script| TransformConfig::new(vrl_script, None));
-    let source_config = SourceConfig {
-        source_id: CLI_INGEST_SOURCE_ID.to_string(),
-        max_num_pipelines_per_indexer: 1,
-        desired_num_pipelines: 1,
-        enabled: true,
-        source_params,
-        transform_config,
-    };
-    run_index_checklist(&config.metastore_uri, &args.index_id, Some(&source_config)).await?;
-    let metastore_uri_resolver = quickwit_metastore_uri_resolver();
-    let metastore = metastore_uri_resolver
-        .resolve(&config.metastore_uri)
-        .await?;
-
-    if args.overwrite {
-        let index_service = IndexService::from_config(config.clone()).await?;
-        index_service.clear_index(&args.index_id).await?;
-    }
-    // The indexing service needs to update its cluster chitchat state so that the control plane is
-    // aware of the running tasks. We thus create a fake cluster to instantiate the indexing service
-    // and avoid impacting potential control plane running on the cluster.
-    let fake_cluster = create_fake_cluster_for_cli().await?;
-    let indexer_config = IndexerConfig {
-        ..Default::default()
-    };
-    start_actor_runtimes(&HashSet::from_iter([QuickwitService::Indexer]))?;
-    let universe = Universe::new();
-    let indexing_server = IndexingService::new(
-        config.node_id.clone(),
-        config.data_dir_path.clone(),
-        indexer_config,
-        Arc::new(fake_cluster),
-        metastore,
-        quickwit_storage_uri_resolver().clone(),
-    )
-    .await?;
-    let (indexing_server_mailbox, indexing_server_handle) =
-        universe.spawn_builder().spawn(indexing_server);
-    let pipeline_id = indexing_server_mailbox
-        .ask_for_res(SpawnPipeline {
-            index_id: args.index_id.clone(),
-            source_config,
-            pipeline_ord: 0,
-        })
-        .await?;
-    let merge_pipeline_handle = indexing_server_mailbox
-        .ask_for_res(DetachMergePipeline {
-            pipeline_id: MergePipelineId::from(&pipeline_id),
-        })
-        .await?;
-    let indexing_pipeline_handle = indexing_server_mailbox
-        .ask_for_res(DetachIndexingPipeline { pipeline_id })
-        .await?;
-
-    let is_stdin_atty = atty::is(atty::Stream::Stdin);
-    if args.input_path_opt.is_none() && is_stdin_atty {
-        let eof_shortcut = match env::consts::OS {
-            "windows" => "CTRL+Z",
-            _ => "CTRL+D",
-        };
-        println!(
-            "Please, enter JSON documents one line at a time.\nEnd your input using {}.",
-            eof_shortcut
-        );
-    }
-    let statistics =
-        start_statistics_reporting_loop(indexing_pipeline_handle, args.input_path_opt.is_none())
-            .await?;
-    merge_pipeline_handle.quit().await;
-    // Shutdown the indexing server.
-    universe
-        .send_exit_with_success(&indexing_server_mailbox)
-        .await?;
-    indexing_server_handle.join().await;
-    if statistics.num_published_splits > 0 {
-        println!(
-            "Now, you can query the index with the following command:\nquickwit index search \
-             --index {} --config ./config/quickwit.yaml --query \"my query\"",
-            args.index_id
-        );
-    }
-
-    if args.clear_cache {
-        println!("Clearing local cache directory...");
-        clear_cache_directory(&config.data_dir_path).await?;
-        println!("{} Local cache directory cleared.", "✔".color(GREEN_COLOR));
-    }
-
-    match statistics.num_invalid_docs {
-        0 => {
-            println!("{} Documents successfully indexed.", "✔".color(GREEN_COLOR));
-            Ok(())
-        }
-        _ => bail!("Failed to ingest all the documents."),
-    }
-}
-
 pub async fn search_index(args: SearchIndexArgs) -> anyhow::Result<SearchResponseRest> {
     let aggs: Option<serde_json::Value> = args
         .aggregation
@@ -1067,85 +783,6 @@ pub async fn search_index_cli(args: SearchIndexArgs) -> anyhow::Result<()> {
     Ok(())
 }
 
-pub async fn merge_cli(args: MergeArgs) -> anyhow::Result<()> {
-    debug!(args=?args, "run-merge-operations");
-    println!("❯ Merging splits locally...");
-    let config = load_quickwit_config(&args.config_uri).await?;
-    run_index_checklist(&config.metastore_uri, &args.index_id, None).await?;
-    let indexer_config = IndexerConfig {
-        ..Default::default()
-    };
-    // The indexing service needs to update its cluster chitchat state so that the control plane is
-    // aware of the running tasks. We thus create a fake cluster to instantiate the indexing service
-    // and avoid impacting potential control plane running on the cluster.
-    let fake_cluster = create_fake_cluster_for_cli().await?;
-    let metastore_uri_resolver = quickwit_metastore_uri_resolver();
-    let metastore = metastore_uri_resolver
-        .resolve(&config.metastore_uri)
-        .await?;
-    let storage_resolver = quickwit_storage_uri_resolver().clone();
-    start_actor_runtimes(&HashSet::from_iter([QuickwitService::Indexer]))?;
-    let indexing_server = IndexingService::new(
-        config.node_id,
-        config.data_dir_path,
-        indexer_config,
-        Arc::new(fake_cluster),
-        metastore,
-        storage_resolver,
-    )
-    .await?;
-    let universe = Universe::new();
-    let (indexing_service_mailbox, indexing_service_handle) =
-        universe.spawn_builder().spawn(indexing_server);
-    let pipeline_id = indexing_service_mailbox
-        .ask_for_res(SpawnPipeline {
-            index_id: args.index_id,
-            source_config: SourceConfig {
-                source_id: args.source_id,
-                max_num_pipelines_per_indexer: 1,
-                desired_num_pipelines: 1,
-                enabled: true,
-                source_params: SourceParams::Vec(VecSourceParams::default()),
-                transform_config: None,
-            },
-            pipeline_ord: 0,
-        })
-        .await?;
-    let pipeline_handle: ActorHandle<MergePipeline> = indexing_service_mailbox
-        .ask_for_res(DetachMergePipeline {
-            pipeline_id: MergePipelineId::from(&pipeline_id),
-        })
-        .await?;
-
-    let mut check_interval = tokio::time::interval(Duration::from_secs(1));
-    loop {
-        check_interval.tick().await;
-
-        let observation = pipeline_handle.observe().await;
-
-        if observation.num_ongoing_merges == 0 {
-            info!("Merge pipeline has no more ongoing merges, Exiting.");
-            break;
-        }
-
-        if observation.obs_type == ObservationType::PostMortem {
-            info!("Merge pipeline has exited, Exiting.");
-            break;
-        }
-    }
-
-    let (pipeline_exit_status, _pipeline_statistics) = pipeline_handle.quit().await;
-    indexing_service_handle.quit().await;
-    if !matches!(
-        pipeline_exit_status,
-        ActorExitStatus::Success | ActorExitStatus::Quit
-    ) {
-        bail!(pipeline_exit_status);
-    }
-    println!("{} Merge successful.", "✔".color(GREEN_COLOR));
-    Ok(())
-}
-
 pub async fn delete_index_cli(args: DeleteIndexArgs) -> anyhow::Result<()> {
     debug!(args=?args, "delete-index");
     println!("❯ Deleting index...");
@@ -1179,69 +816,6 @@ pub async fn delete_index_cli(args: DeleteIndexArgs) -> anyhow::Result<()> {
         return Ok(());
     }
     println!("{} Index successfully deleted.", "✔".color(GREEN_COLOR));
-    Ok(())
-}
-
-pub async fn garbage_collect_index_cli(args: GarbageCollectIndexArgs) -> anyhow::Result<()> {
-    debug!(args=?args, "garbage-collect-index");
-    println!("❯ Garbage collecting index...");
-    quickwit_telemetry::send_telemetry_event(TelemetryEvent::GarbageCollect).await;
-
-    let quickwit_config = load_quickwit_config(&args.config_uri).await?;
-    let index_service = IndexService::from_config(quickwit_config.clone()).await?;
-    let removal_info = index_service
-        .garbage_collect_index(&args.index_id, args.grace_period, args.dry_run)
-        .await?;
-    if removal_info.removed_split_entries.is_empty() && removal_info.failed_split_ids.is_empty() {
-        println!("No dangling files to garbage collect.");
-        return Ok(());
-    }
-
-    if args.dry_run {
-        println!("The following files will be garbage collected.");
-        for file_entry in removal_info.removed_split_entries {
-            println!(" - {}", file_entry.file_name);
-        }
-        return Ok(());
-    }
-
-    if !removal_info.failed_split_ids.is_empty() {
-        println!("The following splits were attempted to be removed, but failed.");
-        for split_id in removal_info.failed_split_ids.iter() {
-            println!(" - {}", split_id);
-        }
-        println!(
-            "{} Splits were unable to be removed.",
-            removal_info.failed_split_ids.len()
-        );
-    }
-
-    let deleted_bytes: u64 = removal_info
-        .removed_split_entries
-        .iter()
-        .map(|entry| entry.file_size_in_bytes)
-        .sum();
-    println!(
-        "{}MB of storage garbage collected.",
-        deleted_bytes / 1_000_000
-    );
-
-    if removal_info.failed_split_ids.is_empty() {
-        println!(
-            "{} Index successfully garbage collected.",
-            "✔".color(GREEN_COLOR)
-        );
-    } else if removal_info.removed_split_entries.is_empty()
-        && !removal_info.failed_split_ids.is_empty()
-    {
-        println!("{} Failed to garbage collect index.", "✘".color(RED_COLOR));
-    } else {
-        println!(
-            "{} Index partially garbage collected.",
-            "✘".color(RED_COLOR)
-        );
-    }
-
     Ok(())
 }
